[DSCP-53] Adding fineer structure to the code
