+++ witness/src/Dscp/Snowdrop/Contracts/Errors.hs
@@ -1,0 +1,17 @@
+
+module Dscp.Snowdrop.Contracts.Errors where
+
+data ContractException
+    = PublicationDoesNotExists
+    | WrongSeller
+    | WrongBuyer
+    | ThisCostsMore
+    | WrongIntialStage
+    | WrongInheritor
+    | WrongPersonToStartTheTrade
+    | SellerUnderpaidTheFees
+    | CancellingOnCorrectSig
+    | NotCancellingOnInCorrectSig
+    | AddressIsAlreadyTaken
+    | ContractDoesNotExist
+    | SelfContractsUnallowed
